+++ spec/rubocop/cop/style/class_and_module_children_spec.rb
@@ -60,6 +60,20 @@ RSpec.describe RuboCop::Cop::Style::ClassAndModuleChildren, :config do
       RUBY
     end
 
+    it 'accepts cbase class name' do
+      expect_no_offenses(<<~RUBY)
+        class ::Foo
+        end
+      RUBY
+    end
+
+    it 'accepts cbase module name' do
+      expect_no_offenses(<<~RUBY)
+        module ::Foo
+        end
+      RUBY
+    end
+
     it 'accepts :: in parent class on inheritance' do
       expect_no_offenses(<<~RUBY)
         class FooClass